--- src/main/java/com/project/springdemo/HelloWorldImpl.java
@@ -1,12 +1,0 @@
-package com.project.springdemo;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class HelloWorldImpl implements HelloWorldService{
-	
-	public void sayHello() {
-		System.out.println("Hello World !!!");
-	}
-
-}
