+++ src/main/java/com/project/springdemo/typesofinjection/InjectedByConsturctorService.java
@@ -1,0 +1,24 @@
+package com.project.springdemo.typesofinjection;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.project.springdemo.services.HelloWorldService;
+
+@Component
+public class InjectedByConsturctorService {
+
+	
+	private HelloWorldService helloWorldService;
+
+	@Autowired
+	public InjectedByConsturctorService(HelloWorldService helloWorldService) {
+		this.helloWorldService = helloWorldService;
+	}
+	
+	public void getMessage() {
+		helloWorldService.getGreetings();
+	}
+	
+	
+}
