+++ src/main/java/com/project/springdemo/typesofinjection/SetterbasedInjection.java
@@ -1,22 +1,21 @@
-package com.project.springdemo;
+package com.project.springdemo.typesofinjection;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-@Component
-public class InjectedByConsturctorService {
+import com.project.springdemo.services.HelloWorldService;
 
-	
+@Component
+public class SetterbasedInjection {
+
 	private HelloWorldService helloWorldService;
 
 	@Autowired
-	public InjectedByConsturctorService(HelloWorldService helloWorldService) {
+	public void setHelloWorldService(HelloWorldService helloWorldService) {
 		this.helloWorldService = helloWorldService;
 	}
-	
+
 	public void getMessage() {
-		helloWorldService.sayHello();
+		helloWorldService.getGreetings();
 	}
-	
-	
 }
